+++ src/main/java/com/devbliss/gpullr/service/UserService.java
@@ -5,6 +5,7 @@ import com.devbliss.gpullr.exception.LoginRequiredException;
 import com.devbliss.gpullr.repository.UserRepository;
 import com.devbliss.gpullr.session.UserSession;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -34,8 +35,17 @@ public class UserService {
     return userRepository.findAll();
   }
 
+  /**
+   * Finds all users that are allowed to login to this application, sorted by username.
+   * 
+   * @return possibly empty list of users
+   */
   public List<User> findAllOrgaMembers() {
-    return userRepository.findByCanLoginIsTrue();
+    return userRepository
+      .findByCanLoginIsTrue()
+      .stream()
+      .sorted((u1, u2) -> u1.username.toLowerCase().compareTo(u2.username.toLowerCase()))
+      .collect(Collectors.toList());
   }
 
   public void requireLogin() throws LoginRequiredException {